Fix: Multiple input file compilation errors
- input_db.cpp: Use uintptr_t instead of DWORD for pointer cast (64-bit safe)
- input_login.cpp: Add <memory> header for std::unique_ptr
- tables.h: Convert strlcpy macro to inline function (returns size_t)

+++ Source/Server/game/src/input_db.cpp
@@ -2400,7 +2400,7 @@ int CInputDB::Analyze(LPDESC d, BYTE bHeader, const char * c_pData)
 
 #ifdef ENABLE_EVENT_MANAGER
 		case HEADER_DG_UPDATE_EVENT_STATUS:
-			CEventManager::Instance().UpdateEventStatus((DWORD)c_pData);
+			CEventManager::Instance().UpdateEventStatus((uintptr_t)c_pData);
 
 		case HEADER_DG_EVENT_NOTIFICATION:
 			EventNotification(c_pData);
